+++ Changelog.md
@@ -18,6 +18,7 @@
 - Opensearch Upgrade to 2.19.0
 - Properly fetch GLOBAL_TENANT_DOMAIN value from parameter store
 - Prevent customer changes to environment during upgrades via maintenance page
+- Refactor os-bootstrap code, remove wait-for-os-health initContainer
 
 _Changes:_
 
@@ -39,7 +40,9 @@ _Changes:_
 - [X] PDO-6878 Move PingFederate IK upload scripts from ping-cloud-base to ping-cloud-tools repository
 - [X] PDO-9600 Properly fetch GLOBAL_TENANT_DOMAIN value from parameter store
 - [X] PDO-9798 ingress-nginx-private autoscaling min/max replicas
+- [X] PDO-9972 [Logstash] OS bootstrap code refactor
 - [X] PDO-10087 Prevent customer changes to environment during upgrades via maintenance page
+
 
 ### 2.1.1